Use `priv_file` and `priv_dir` cowboy_static options
Before, we computed the absolute path of a file inside an Erlang
application, and used the `file` and `dir` options of cowboy_static.

Now, we also handle the case where the file in the application (owning
the specified module) priv directory. In this case, we rely on the
`priv_file` and `priv_dir` options: cowboy_static will automatically
locate the application's `priv` directory and Do The Right Thing.

In the future, this will allow us to serve static files from Erlang
application .ez archives.

[#138338155]