Ensure that cowboy_req:peer return values will not crash
Since the type spec for cowboy_req:peer indicates that a two-tuple of {undefined, Req} can be returned
we should handle this as `badmatch` could be the result otherwise.

I also used the returned value of Req as specified in the cowboy docs:

https://ninenines.eu/docs/en/cowboy/1.0/guide/req/

I realize this is pedantic and unnecessary in this case but it ensures that we can handle changes
in cowboy should the API change in the future.